refactor(labs): added test to improved labs search
Added a test to prove search labs by code functionality

fix #1971
